Consolidate Button and IconButton into /ui/button.tsx
- Merge custom button with shadcn patterns (data-slot, cleaner structure)
- Move IconButton to use CVA for variants/sizes like Button
- Remove redundant shadcn variant aliases
- Update all imports to use @/app/components/ui/button
- Delete old app/components/button.tsx
