+++ app/components/ui/button.tsx
@@ -1,62 +1,209 @@
+"use client"
+
 import * as React from "react"
 import { Slot } from "@radix-ui/react-slot"
 import { cva, type VariantProps } from "class-variance-authority"
+import Link from "next/link"
 
 import { cn } from "@/app/utils"
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+} from "@/app/components/ui/tooltip"
 
 const buttonVariants = cva(
-  "cursor-pointer inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-all disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 shrink-0 [&_svg]:shrink-0 outline-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive",
+  "cursor-pointer inline-flex items-center justify-center gap-2 whitespace-nowrap font-medium transition-all duration-300 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 shrink-0 [&_svg]:shrink-0 outline-none focus-visible:ring-ring/50 focus-visible:ring-[3px] relative overflow-hidden rounded-full before:absolute before:inset-0 before:transition-all before:duration-300 before:rounded-full before:scale-75 hover:before:scale-100",
   {
     variants: {
       variant: {
-        default: "bg-primary text-primary-foreground hover:bg-primary/90",
-        destructive:
-          "bg-destructive text-white hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 dark:bg-destructive/60",
-        outline:
-          "border bg-background shadow-xs hover:bg-accent hover:text-accent-foreground dark:bg-input/30 dark:border-input dark:hover:bg-input/50",
+        primary:
+          "text-white bg-neutral-900 hover:before:bg-white/20 dark:text-neutral-900 dark:bg-white dark:hover:before:bg-neutral-900/20",
         secondary:
-          "bg-secondary text-secondary-foreground hover:bg-secondary/80",
-        ghost:
-          "hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50",
-        link: "text-primary underline-offset-4 hover:underline",
+          "text-neutral-900 bg-neutral-900/10 hover:before:bg-neutral-900/20 dark:text-white dark:bg-white/10 dark:hover:before:bg-white/20",
+        tertiary:
+          "text-neutral-700 border border-neutral-900/20 hover:text-neutral-900 hover:before:bg-neutral-900/10 hover:border-neutral-900/30 dark:text-neutral-300 dark:border-white/20 dark:hover:text-white dark:hover:before:bg-white/10 dark:hover:border-white/30",
+        text: "text-neutral-700 hover:text-neutral-900 hover:before:bg-neutral-900/10 dark:text-neutral-300 dark:hover:text-white dark:hover:before:bg-white/10",
+        destructive:
+          "bg-destructive text-white hover:bg-destructive/90 dark:bg-destructive/60",
+        link: "text-primary underline-offset-4 hover:underline before:hidden",
       },
       size: {
-        default: "h-9 px-4 py-2 has-[>svg]:px-3",
-        sm: "h-8 rounded-md gap-1.5 px-3 has-[>svg]:px-2.5",
-        lg: "h-10 rounded-md px-6 has-[>svg]:px-4",
-        icon: "size-9",
-        "icon-sm": "size-8",
-        "icon-lg": "size-10",
+        large: "h-10 px-5 text-sm",
+        medium: "h-8 px-4 text-sm",
+        small: "h-7 px-3 text-xs",
       },
     },
     defaultVariants: {
-      variant: "default",
-      size: "default",
+      variant: "primary",
+      size: "large",
     },
   }
 )
 
-function Button({
-  className,
-  variant = "default",
-  size = "default",
-  asChild = false,
-  ...props
-}: React.ComponentProps<"button"> &
-  VariantProps<typeof buttonVariants> & {
-    asChild?: boolean
-  }) {
-  const Comp = asChild ? Slot : "button"
-
-  return (
-    <Comp
-      data-slot="button"
-      data-variant={variant}
-      data-size={size}
-      className={cn(buttonVariants({ variant, size, className }))}
-      {...props}
-    />
-  )
+interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  href?: string
+  asChild?: boolean
+  absolute?: boolean
+  as?: keyof JSX.IntrinsicElements
 }
 
-export { Button, buttonVariants }
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  (
+    {
+      className,
+      variant,
+      size,
+      href,
+      asChild = false,
+      absolute,
+      as,
+      children,
+      ...props
+    },
+    ref
+  ) => {
+    const position = absolute === true ? "absolute" : "relative"
+    const classes = cn(buttonVariants({ variant, size, className }), position)
+
+    const content = (
+      <span className="relative z-10 flex items-center justify-center gap-2">
+        {children}
+      </span>
+    )
+
+    if (href) {
+      return (
+        <Link href={href} className={classes} data-slot="button">
+          {content}
+        </Link>
+      )
+    }
+
+    if (asChild) {
+      return (
+        <Slot
+          className={classes}
+          ref={ref as React.Ref<HTMLElement>}
+          data-slot="button"
+          {...props}
+        >
+          {children}
+        </Slot>
+      )
+    }
+
+    if (as) {
+      const Tag = as as keyof JSX.IntrinsicElements
+      return (
+        <Tag className={classes} data-slot="button" {...(props as object)}>
+          {content}
+        </Tag>
+      )
+    }
+
+    return (
+      <button className={classes} ref={ref} data-slot="button" {...props}>
+        {content}
+      </button>
+    )
+  }
+)
+Button.displayName = "Button"
+
+// IconButton variants
+const iconButtonVariants = cva(
+  "cursor-pointer overflow-hidden font-medium transition-colors duration-300 rounded-full flex items-center justify-center before:absolute before:inset-0 before:transition-all before:duration-300 before:rounded-full before:scale-75 hover:before:scale-100 disabled:opacity-50 disabled:before:bg-transparent disabled:hover:before:bg-transparent",
+  {
+    variants: {
+      variant: {
+        primary:
+          "text-white bg-neutral-900 hover:before:bg-white/20 dark:text-neutral-900 dark:bg-white dark:hover:before:bg-neutral-900/20",
+        secondary:
+          "text-neutral-900 bg-neutral-900/10 hover:before:bg-neutral-900/20 dark:text-white dark:bg-white/10 dark:hover:before:bg-white/20",
+        tertiary:
+          "text-neutral-700 border border-neutral-900/20 hover:text-neutral-900 hover:before:bg-neutral-900/10 hover:border-neutral-900/30 dark:text-neutral-300 dark:border-white/20 dark:hover:text-white dark:hover:before:bg-white/10 dark:hover:border-white/30",
+        text: "text-neutral-700 hover:text-neutral-900 hover:before:bg-neutral-900/10 dark:text-neutral-300 dark:hover:text-white dark:hover:before:bg-white/10",
+      },
+      size: {
+        large: "h-10 w-10",
+        medium: "h-8 w-8",
+        small: "h-7 w-7",
+      },
+    },
+    defaultVariants: {
+      variant: "primary",
+      size: "large",
+    },
+  }
+)
+
+interface IconButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof iconButtonVariants> {
+  label: string
+  href?: string
+  absolute?: boolean
+}
+
+const IconButton = React.forwardRef<HTMLButtonElement, IconButtonProps>(
+  (
+    {
+      className,
+      variant,
+      size,
+      label,
+      href,
+      absolute,
+      children,
+      ...props
+    },
+    ref
+  ) => {
+    const position = absolute === true ? "absolute" : "relative"
+    const classes = cn(iconButtonVariants({ variant, size, className }), position)
+
+    const content = (
+      <span className="relative z-10 flex items-center justify-center">
+        {children}
+      </span>
+    )
+
+    if (href) {
+      return (
+        <Tooltip>
+          <TooltipTrigger asChild>
+            <Link href={href} className={classes} data-slot="icon-button">
+              {content}
+            </Link>
+          </TooltipTrigger>
+          <TooltipContent>{label}</TooltipContent>
+        </Tooltip>
+      )
+    }
+
+    return (
+      <Tooltip>
+        <TooltipTrigger asChild>
+          <button
+            type={props.type}
+            aria-label={label}
+            className={classes}
+            ref={ref}
+            data-slot="icon-button"
+            {...props}
+          >
+            {content}
+          </button>
+        </TooltipTrigger>
+        <TooltipContent>{label}</TooltipContent>
+      </Tooltip>
+    )
+  }
+)
+IconButton.displayName = "IconButton"
+
+export { Button, IconButton, buttonVariants, iconButtonVariants }
+export type { ButtonProps, IconButtonProps }
