+++ app/projects/[slug]/error.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Animate from "@/app/components/animate";
-import Button from "@/app/components/button";
+import { Button } from "@/app/components/ui/button";
 import { Heading, Text } from "@/app/components/text";
 import { useEffect } from "react";
 import { IconAlertCircle } from "@tabler/icons-react";
